Move catalog action into catalog slice
