+++ assets/js/components/ChecksCatalog/ChecksCatalogPage.jsx
@@ -4,7 +4,7 @@ import { useSelector, useDispatch } from 'react-redux';
 import { pickBy } from 'lodash';
 
 import { getCatalog } from '@state/selectors/catalog';
-import { updateCatalog } from '@state/actions/catalog';
+import { updateCatalog } from '@state/catalog';
 import { OPTION_ALL } from '@components/Select';
 
 import ChecksCatalog from './ChecksCatalog';
