+++ assets/js/components/ClusterSettingsPage/ClusterSettingsPage.jsx
@@ -10,7 +10,7 @@ import {
   getClusterSelectedChecks,
   getClusterHosts,
 } from '@state/selectors/cluster';
-import { updateCatalog } from '@state/actions/catalog';
+import { updateCatalog } from '@state/catalog';
 import { getCatalog } from '@state/selectors/catalog';
 import { isSaving } from '@state/selectors/checksSelection';
 import { executionRequested } from '@state/actions/lastExecutions';
