+++ assets/js/components/HostDetails/HostDetailsPage.jsx
@@ -14,7 +14,7 @@ import { getLastExecution } from '@state/selectors/lastExecutions';
 
 import { getHost, getHostSelectedChecks } from '@state/selectors/host';
 import { isSaving } from '@state/selectors/checksSelection';
-import { updateCatalog } from '@state/actions/catalog';
+import { updateCatalog } from '@state/catalog';
 import {
   updateLastExecution,
   hostExecutionRequested,
