--- assets/js/state/actions/catalog.js
@@ -1,6 +1,0 @@
-export const UPDATE_CATALOG = 'UPDATE_CATALOG';
-
-export const updateCatalog = (env = {}) => ({
-  type: UPDATE_CATALOG,
-  payload: env,
-});
