+++ assets/js/state/sagas/catalog.js
@@ -1,11 +1,11 @@
 import { put, call, takeEvery } from 'redux-saga/effects';
 import { getCatalog } from '@lib/api/checks';
-import { UPDATE_CATALOG } from '@state/actions/catalog';
 
 import {
   setCatalogLoading,
   setCatalogData,
   setCatalogError,
+  UPDATE_CATALOG,
 } from '@state/catalog';
 
 export function* updateCatalog({ payload }) {
